Updated MergeByEnrichProcessor to new API
